refactor: Receipt
- 복잡한 CSS로 인해 유지보수성이 극악이었던 Receipt 컴포넌트를 개선합니다.
- 적절한 단위로 함수화를 통해 코드를 분할하여 유지보수성을 개선합니다.
- 디자인 원안에 대해 오차가 발생했던 부분을 수정하여 최대한 디자인 원안에 일치시킵니다.
