+++ src/components/common/receipt/Receipt.stories.tsx
